requirements.txt
streamlit
groq
crewai
crewai-tools
python-dotenv
qdrant-client
youtube-transcript-api
google-genai